# openapi.yaml
openapi: 3.1.0
info:
  title: Домовой — API приложения дома в MAX
  version: 1.0.0
  description: 'API мини-приложения и бота MAX «Домовой»: вход по QR-коду квитанции, обращения в УК, начисления,
    счётчики, лента и опросы дома, раздел председателя, кабинет УК, подбор дома.


    **Как войти для проверки.** Без квитанции и без MAX — ролью демо-дома: `POST /api/demo/roles/{key}/take` (key:
    chairman, owner12, member12, owner45, owner78, newcomer30) вернёт `token`; дальше заголовок `Authorization:
    Bearer <token>`. Кабинет УК — `POST /api/dispatcher/login` с логином демо-УК (пароль — на служебном слайде презентации),
    тоже `token`.


    **Ошибки** — JSON `{ "error": "код", "message": "текст для человека" }`.


    **Модель доступа.** Квитанция — заявка, а не пропуск: новая квартира приходит со статусом `pending`. Уровень
    0 (pending) — свои начисления, счётчики и обращения в УК сразу. Уровень 1 (active) — соседи, лента, опросы.
    Уровень 2 — дом целиком, только председателю. Обращение удалить нельзя.


    Кабинет оператора (`/api/admin/*`) — внутренний и здесь не описан.'
servers:
- url: https://domovoy.devcore.com.ru
  description: Бой
- url: http://localhost:3000
  description: Локальный запуск (docker compose)
...
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  /api/auth/max:
    post:
      tags:
      - Вход
      summary: Вход через MAX
      description: 'По подписи initData MAX. Если у аккаунта MAX ещё нет квартиры — needs_receipt: нужно отсканировать
        квитанцию.'
      operationId: post_api_auth_max
      parameters:
      - $ref: '#/components/parameters/MaxInitData'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                initData:
                  type:
                  - string
                  - 'null'
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  name:
                    type:
                    - string
                    - 'null'
                  startParam: {}
                  token:
                    type: string
                required:
                - status
        '401':
          description: Нужен вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
...
  /api/pick/favorites:
    get:
      tags:
      - Подбор дома
      summary: Мои дома
      description: Отмеченные ♥.
      operationId: get_api_pick_favorites
      parameters:
      - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  houses:
                    type: array
                    items:
                      type: object
                      properties:
                        houseKey:
                          type:
                          - string
                          - 'null'
                        address:
                          type:
                          - string
                          - 'null'
                        rating: {}
                        reviews:
                          type:
                          - number
                          - 'null'
                      required:
                      - houseKey
                      - address
                      - rating
                      - reviews
                required:
                - houses
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
  /api/pick/house/{houseKey}:
    get:
      tags:
      - Подбор дома
      summary: Паспорт дома
      description: Год, этажность, УК, отзывы без авторов, сводки выше порогов.
      operationId: get_api_pick_house_houseKey
      parameters:
      - name: houseKey
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  photo:
                    description: Только у ЖК заглушки (ключ mock:…) — рендер застройщика с подписью.
                    type:
                    - object
                    - 'null'
                    properties:
                      url:
                        type: string
                      credit:
                        type: string
                  houseKey:
                    type:
                    - string
                    - 'null'
                  address:
                    type:
                    - string
                    - 'null'
                  lat:
                    type:
                    - number
                    - 'null'
                  lon:
                    type:
                    - number
                    - 'null'
                  passport:
                    type: object
                    properties:
                      builtYear: {}
                      floors: {}
                      flats:
                        type:
                        - number
                        - 'null'
                      entrances: {}
                      elevators: {}
                      wallMaterial: {}
                      gas: {}
                      emergency: {}
                    required:
                    - builtYear
                    - floors
                    - flats
                    - entrances
                    - elevators
                    - wallMaterial
                    - gas
                    - emergency
                  management:
                    type: object
                    properties:
                      form:
                        type:
                        - string
                        - 'null'
                      orgName: {}
                      license: {}
                      phone: {}
                      hasChairman:
                        type:
                        - boolean
                        - 'null'
                    required:
                    - form
                    - orgName
                    - license
                    - phone
                    - hasChairman
                  complaints:
                    type: object
                    properties:
                      enough:
                        type:
                        - boolean
                        - 'null'
                      speed: {}
                      topCategory: {}
                    required:
                    - enough
                    - speed
                    - topCategory
                  payment: {}
                  reviews:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        stars:
                          type:
                          - number
                          - object
                        overall:
                          type:
                          - number
                          - 'null'
                        pros:
                          type:
                          - string
                          - 'null'
                        cons: {}
                        createdAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        updatedAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        mine:
                          type:
                          - boolean
                          - 'null'
                        hiddenReason: {}
                        flat:
                          type:
                          - string
                          - 'null'
                        name:
                          type:
                          - string
                          - 'null'
                        plus:
                          type:
                          - string
                          - 'null'
                        minus:
                          type:
                          - string
                          - 'null'
                      required:
                      - stars
                  summary:
                    type: object
                    properties:
                      rating:
                        type:
                        - number
                        - 'null'
                      count:
                        type:
                        - number
                        - 'null'
                      aspects:
                        type: object
                        properties:
                          uk:
                            type:
                            - number
                            - 'null'
                          clean:
                            type:
                            - number
                            - 'null'
                          neighbors:
                            type:
                            - number
                            - 'null'
                          quiet:
                            type:
                            - number
                            - 'null'
                          yard:
                            type:
                            - number
                            - 'null'
                        required:
                        - uk
                        - clean
                        - neighbors
                        - quiet
                        - yard
                    required:
                    - rating
                    - count
                    - aspects
                  near: {}
                  listings:
                    type: array
                    items:
                      type: object
                      properties:
                        site:
                          type:
                          - string
                          - 'null'
                        label:
                          type:
                          - string
                          - 'null'
                        url:
                          type:
                          - string
                          - 'null'
                      required:
                      - site
                      - label
                      - url
                  favorite:
                    type:
                    - boolean
                    - 'null'
                  canReview:
                    type:
                    - boolean
                    - 'null'
                  kind:
                    type:
                    - string
                    - 'null'
                  key:
                    type: string
                  name:
                    type:
                    - string
                    - 'null'
                  microdistrict:
                    type:
                    - string
                    - 'null'
                  district:
                    type:
                    - string
                    - 'null'
                  developer:
                    type:
                    - string
                    - 'null'
                  priceFrom:
                    type:
                    - number
                    - 'null'
                  grocery:
                    type:
                    - boolean
                    - 'null'
                  blurb:
                    type:
                    - string
                    - 'null'
                  tags:
                    type: array
                    items:
                      type: string
                  rating:
                    type:
                    - number
                    - 'null'
                  reviewCount:
                    type:
                    - number
                    - 'null'
                  similar:
                    type: array
                    items: {}
                required:
                - address
                - lat
                - lon
                - reviews
                - listings
        '401':
          description: Нужен вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
  /api/pick/house/{houseKey}/review:
    post:
      tags:
      - Подбор дома
      summary: Отзыв о доме
      description: Только житель с подтверждённым доступом к этому дому. Автор наружу не уходит.
      operationId: post_api_pick_house_houseKey_review
      parameters:
      - name: houseKey
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stars:
                  type: object
                  properties:
                    uk:
                      type:
                      - number
                      - 'null'
                    clean:
                      type:
                      - number
                      - 'null'
                    neighbors:
                      type:
                      - number
                      - 'null'
                    quiet:
                      type:
                      - number
                      - 'null'
                    yard:
                      type:
                      - number
                      - 'null'
                  required:
                  - uk
                  - clean
                  - neighbors
                  - quiet
                  - yard
                pros:
                  type:
                  - string
                  - 'null'
              required:
              - stars
              - pros
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type:
                    - boolean
                    - 'null'
                required:
                - ok
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pick/mock-photo/{slug}:
    get:
      tags:
      - Подбор дома
      summary: Фото ЖК заглушки
      description: Рендер застройщика (заглушка MVP, docs/mock-complexes.md). Открыто без входа — картинку грузит браузер без
        заголовка MAX. Кэш на сутки.
      operationId: get_api_pick_mock_photo_slug
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Картинка
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pick/districts:
    get:
      tags:
      - Подбор дома
      summary: Районы города для отдалённой карты
      description: Граница района (GeoJSON, OpenStreetMap), точка подписи, рамка и число ЖК. Пусто, пока не загружена заглушка ЖК.
      operationId: get_api_pick_districts
      parameters:
      - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  districts:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        count:
                          type: number
                        bbox:
                          type: object
                          properties:
                            south:
                              type: number
                            north:
                              type: number
                            west:
                              type: number
                            east:
                              type: number
                          required:
                          - south
                          - north
                          - west
                          - east
                        label:
                          type: object
                          properties:
                            lat:
                              type: number
                            lon:
                              type: number
                          required:
                          - lat
                          - lon
                        focus:
                          description: Середина ЖК района — туда смотрит карта, если район не влезает в экран
                          type:
                          - object
                          - 'null'
                          properties:
                            lat:
                              type: number
                            lon:
                              type: number
                        geometry:
                          type: object
                          properties:
                            type:
                              type: string
                              enum:
                              - Polygon
                              - MultiPolygon
                            coordinates:
                              type: array
                          required:
                          - type
                          - coordinates
                      required:
                      - name
                      - count
                      - bbox
                      - label
                      - focus
                      - geometry
                required:
                - districts
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
  /api/pick/houses:
    get:
      tags:
      - Подбор дома
      summary: Дома на карте
      description: Дома или ЖК в видимой области карты (bbox, zoom). Открыт гостю MAX без квитанции.
      operationId: get_api_pick_houses
      parameters:
      - name: bbox
        in: query
        required: false
        schema:
          type: string
      - name: zoom
        in: query
        required: false
        schema:
          type: string
      - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind:
                    type:
                    - string
                    - 'null'
                  houses:
                    type: array
                    items:
                      type: object
                      properties:
                        houseKey:
                          type:
                          - string
                          - 'null'
                        lat:
                          type:
                          - number
                          - 'null'
                        lon:
                          type:
                          - number
                          - 'null'
                        address:
                          type:
                          - string
                          - 'null'
                        builtYear: {}
                        floors: {}
                        rating: {}
                        reviews:
                          type:
                          - number
                          - 'null'
                      required:
                      - houseKey
                      - lat
                      - lon
                      - address
                      - builtYear
                      - floors
                      - rating
                      - reviews
                  complexes:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        name:
                          type:
                          - string
                          - 'null'
                        address:
                          type:
                          - string
                          - 'null'
                        microdistrict:
                          type:
                          - string
                          - 'null'
                        district:
                          type:
                          - string
                          - 'null'
                        lat:
                          type:
                          - number
                          - 'null'
                        lon:
                          type:
                          - number
                          - 'null'
                        developer:
                          type:
                          - string
                          - 'null'
                        priceFrom:
                          type:
                          - number
                          - 'null'
                        grocery:
                          type:
                          - boolean
                          - 'null'
                        rating:
                          type:
                          - number
                          - 'null'
                        reviews:
                          type:
                          - number
                          - 'null'
                        quote:
                          type:
                          - string
                          - 'null'
                      required:
                      - key
                      - name
                      - address
                      - microdistrict
                      - district
                      - lat
                      - lon
                      - developer
                      - priceFrom
                      - grocery
                      - rating
                      - reviews
                      - quote
                required:
                - kind
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
  /api/pick/prompt:
    get:
      tags:
      - Подбор дома
      summary: Просьба оценить свой дом
      operationId: get_api_pick_prompt
      parameters:
      - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type:
                    - object
                    - 'null'
                    properties:
                      houseKey:
                        type:
                        - string
                        - 'null'
                      address:
                        type:
                        - string
                        - 'null'
                      reviews:
                        type:
                        - number
                        - 'null'
                    required:
                    - houseKey
                    - address
                    - reviews
                required:
                - prompt
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
  /api/pick/prompt/{houseKey}/dismiss:
    post:
      tags:
      - Подбор дома
      summary: Не просить оценить
      operationId: post_api_pick_prompt_houseKey_dismiss
      parameters:
      - name: houseKey
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type:
                    - boolean
                    - 'null'
                required:
                - ok
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен сессии из /api/auth/qr, /api/auth/max, /api/demo/roles/{key}/take или /api/dispatcher/login
    cookieAuth:
      type: apiKey
      in: cookie
      name: zd_session
      description: Та же сессия в cookie — её ставит сервер при входе
  parameters:
    MaxInitData:
      name: X-Max-Init-Data
      in: header
      required: false
      schema:
        type: string
      description: Подписанная строка initData мини-приложения MAX (WebApp.initData). Проверяется подпись токеном
        бота.
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Машинный код ошибки
        message:
          type: string
          description: Понятный человеку текст
  responses:
    Unauthorized:
      description: Нужен вход
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'